requirements.txt
requests
pyyaml
discord.py
